requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
filelock==3.15.1
gunicorn==22.0.0
identify==2.5.36
idna==3.7
inflection==0.5.1
isort==5.13.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
